move UploadResult.java
