+++ ttnhabbridge/src/main/java/nl/sikken/bertrik/hab/habitat/HabitatUploader.java
@@ -18,6 +18,7 @@ import javax.xml.bind.DatatypeConverter;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
+import nl.sikken.bertrik.hab.UploadResult;
 import nl.sikken.bertrik.hab.habitat.docs.ListenerInformationDoc;
 import nl.sikken.bertrik.hab.habitat.docs.ListenerTelemetryDoc;
 import nl.sikken.bertrik.hab.habitat.docs.PayloadTelemetryDoc;
