+++ ttnhabbridge/src/main/java/nl/sikken/bertrik/hab/habitat/IHabitatRestApi.java
@@ -1,5 +1,6 @@
 package nl.sikken.bertrik.hab.habitat;
 
+import nl.sikken.bertrik.hab.UploadResult;
 import retrofit2.Call;
 import retrofit2.http.Body;
 import retrofit2.http.GET;
